Fix #5 로그인 관련 코드 수정
카카오 로그인시 토큰 저장명 변경
카카오 로그인 필요없는 버튼 삭제
